# cmd/management/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Source-of-truth API for managing GO Feature Flag flagsets, flags, teams, and audit history.",
        "title": "GO Feature Flag Management API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/callback": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "OIDC callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OIDC authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OIDC state",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
...
        "/flags/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flags"
                ],
                "summary": "Get a flag with its current version payload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Flag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flags"
                ],
                "summary": "Replace flag payload (creates a new version)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "new payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateFlagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Flag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "flags"
                ],
                "summary": "Soft-delete a flag (audited; versions remain)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/flags/{id}/disable": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flags"
                ],
                "summary": "Toggle the disabled state of a flag (no new version)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flag id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "disabled state",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DisableFlagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/flagsets/{flagsetId}/flags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flags"
                ],
                "summary": "List flags of a flagset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "flagsetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name contains filter",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "filter disabled flags",
                        "name": "disabled",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "items per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.PaginatedResponse-model_FlagListItem"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flags"
                ],
                "summary": "Create a flag in a flagset (creates version 1)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "flagset id",
                        "name": "flagsetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "flag payload (GOFF DTO)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateFlagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Flag"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.CreateFlagRequest": {
            "type": "object"
        },
...
        "model.DisableFlagRequest": {
            "type": "object",
            "properties": {
                "disabled": {
                    "type": "boolean"
                }
            }
        },
        "model.Flag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "currentVersion": {
                    "type": "integer"
                },
                "currentVersionId": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "flagsetId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.FlagListItem": {
            "type": "object",
            "properties": {
                "currentVersion": {
                    "type": "integer"
                },
                "disabled": {
                    "type": "boolean"
                },
                "flagsetId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.PaginatedResponse-model_FlagListItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FlagListItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "model.UpdateFlagRequest": {
            "type": "object"
        },
...
    }
}